Move setting of projects from command-line argument to code
Previously, the current project in Maya was set upon launch, passing the -proj flag along with an absolute path to the project. This resulted in a cyclic dependency between the Application Definition and environment, as the project directory dependend on the name of the application directory found in the application definition, and the definition itself being dependent on the environment. Now the application definition includes information on where to store working files, but doesn't know upfront where those files will be stored as the directory is created further down the pipeline.

Small detail, but wanted to fully document for you and my future self!