Add `bpy.context` wrapper
Because the `bpy.context` is not always properly populated when queried
from within a Qt app, a wrapper is added. Instead of directly using
`bpy` you can `from avalon.blender import bpy`. All needed overrides can
now be handled there and everything else is simply passed through from
the original `bpy`.

Hopefully this is a (fairly) short-term remedy and a proper solution
will be found in the near future. 8)