Speed up getting assets for asset switcher
For every asset a database call was made. If you have lots of assets
this can really add up and take quite some time (for me roughly 30
seconds for about 2000 assets).
This change does only 2 database calls and then does the filtering
afterwards. This reduces the time taken to around 3 seconds for about
2000 assets.